+++ XIVAuras/Auras/AuraIcon.cs
@@ -69,7 +69,12 @@ namespace XIVAuras.Auras
             }
 
             bool visible = this.VisibilityConfig.IsVisible(parentVisible);
-            bool triggered = this.TriggerConfig.IsTriggered(this.Preview, out DataSource data) && visible;
+            if (!visible && !this.Preview)
+            {
+                return;
+            }
+
+            bool triggered = this.TriggerConfig.IsTriggered(this.Preview, out DataSource data);
             IconStyleConfig style = this.StyleConditions.GetStyle(data) ?? this.IconStyleConfig;
 
             Vector2 localPos = pos + style.Position;
@@ -90,6 +95,11 @@ namespace XIVAuras.Auras
                             localPos = ImGui.GetWindowPos();
                             style.Position = localPos - pos;
                         }
+                    }
+
+                    if (style.IconOption == 2)
+                    {
+                        return;
                     }
                     
                     bool desaturate = style.DesaturateIcon;
@@ -114,22 +124,32 @@ namespace XIVAuras.Auras
                         }
                     }
 
-                    if (style.IconOption != 2)
+                    if (style.ShowProgressSwipe)
                     {
-                        if (style.ShowProgressSwipe)
+                        if (style.GcdSwipe && (data.Value == 0 || data.MaxValue == 0 || style.GcdSwipeOnly))
                         {
-                            this.DrawProgressSwipe(style, localPos, size, data.Value, data.MaxValue, alpha, drawList);
+                            SpellHelpers.GetGCDInfo(out var recastInfo);
+                            DrawProgressSwipe(style, localPos, size, recastInfo.RecastTime - recastInfo.RecastTimeElapsed, recastInfo.RecastTime, alpha, drawList);
                         }
+                        else
+                        {
+                            DrawProgressSwipe(style, localPos, size, data.Value, data.MaxValue, alpha, drawList);
+                        }
+                    }
 
-                        if (style.ShowBorder)
+                    if (style.ShowBorder)
+                    {
+                        for (int i = 0; i < style.BorderThickness; i++)
                         {
-                            for (int i = 0; i < style.BorderThickness; i++)
-                            {
-                                Vector2 offset = new Vector2(i, i);
-                                Vector4 color = style.BorderColor.Vector.AddTransparency(alpha);
-                                drawList.AddRect(localPos + offset, localPos + size - offset, ImGui.ColorConvertFloat4ToU32(color));
-                            }
+                            Vector2 offset = new Vector2(i, i);
+                            Vector4 color = style.BorderColor.Vector.AddTransparency(alpha);
+                            drawList.AddRect(localPos + offset, localPos + size - offset, ImGui.ColorConvertFloat4ToU32(color));
                         }
+                    }
+
+                    if (style.Glow)
+                    {
+                        this.DrawIconGlow(localPos, size, style.GlowThickness, style.GlowSegments, style.GlowSpeed, style.GlowColor, style.GlowColor2, drawList);
                     }
                 });
             }
@@ -160,7 +180,7 @@ namespace XIVAuras.Auras
             this.LastFrameWasPreview = this.Preview;
         }
 
-        private void DrawProgressSwipe(
+        private static void DrawProgressSwipe(
             IconStyleConfig style,
             Vector2 pos,
             Vector2 size,
@@ -198,6 +218,25 @@ namespace XIVAuras.Auras
 
                 ImGui.PopClipRect();
             }
+        }
+
+        private void DrawIconGlow(Vector2 pos, Vector2 size, int thickness, int segments, float speed, ConfigColor col1, ConfigColor col2, ImDrawListPtr drawList)
+        {
+            speed = Math.Abs(speed);
+            int mod = speed == 0 ? 1 : (int)(250 / speed);
+            float prog = (float)(DateTimeOffset.Now.ToUnixTimeMilliseconds() % mod) / mod;
+
+            float offset = thickness / 2 + thickness % 2;
+            Vector2 pad = new Vector2(offset);
+            Vector2 c1 = new Vector2(pos.X, pos.Y);
+            Vector2 c2 = new Vector2(pos.X + size.X, pos.Y);
+            Vector2 c3 = new Vector2(pos.X + size.X, pos.Y + size.Y);
+            Vector2 c4 = new Vector2(pos.X, pos.Y + size.Y);
+
+            DrawHelpers.DrawSegmentedLineHorizontal(drawList, c1, size.X, thickness, prog, segments, col1, col2);
+            DrawHelpers.DrawSegmentedLineVertical(drawList, c2.AddX(-thickness), thickness, size.Y, prog, segments, col1, col2);
+            DrawHelpers.DrawSegmentedLineHorizontal(drawList, c3.AddY(-thickness), -size.X, thickness, prog, segments, col1, col2);
+            DrawHelpers.DrawSegmentedLineVertical(drawList, c4, thickness, -size.Y, prog, segments, col1, col2);
         }
 
         public static AuraIcon GetDefaultAuraIcon(string name)