+++ XIVAuras/Config/LabelStyleConfig.cs
@@ -10,6 +10,7 @@ namespace XIVAuras.Config
     public class LabelStyleConfig : IConfigPage
     {
         [JsonIgnore] private string[] _anchorOptions = Enum.GetNames(typeof(DrawAnchor));
+        [JsonIgnore] private string[] _roundingOptions = new string[] { "Floor", "Ceiling", "Round" };
 
         public string Name => "Text";
 
@@ -22,6 +23,7 @@ namespace XIVAuras.Config
         public ConfigColor TextColor = new ConfigColor(1, 1, 1, 1);
         public bool ShowOutline = true;
         public ConfigColor OutlineColor = new ConfigColor(0, 0, 0, 1);
+        public int Rounding = 0;
 
         public LabelStyleConfig() : this("") { }
 
@@ -36,12 +38,13 @@ namespace XIVAuras.Config
         {
             if (ImGui.BeginChild("##LabelStyleConfig", new Vector2(size.X, size.Y), true))
             {
-                ImGui.InputTextWithHint("Text Format", "[value] or [stacks]", ref this.TextFormat, 64);
+                ImGui.InputTextWithHint("Text Format", "Hover for Formatting Info", ref this.TextFormat, 64);
                 if (ImGui.IsItemHovered())
                 {
                     ImGui.SetTooltip(Utils.GetTagsTooltip(DataSource.TextTags));
                 }
 
+                ImGui.Combo("Number Format", ref this.Rounding, _roundingOptions, _roundingOptions.Length);
                 ImGui.DragFloat2("Position", ref this.Position);
                 ImGui.Combo("Parent Anchor", ref Unsafe.As<DrawAnchor, int>(ref this.ParentAnchor), _anchorOptions, _anchorOptions.Length);
                 ImGui.Combo("Text Align", ref Unsafe.As<DrawAnchor, int>(ref this.TextAlign), _anchorOptions, _anchorOptions.Length);