+++ XIVAuras/Config/TriggerOptions.cs
@@ -16,22 +16,5 @@ namespace XIVAuras.Config
         public abstract TriggerSource Source { get; }
         public abstract bool IsTriggered(bool preview, out DataSource data);
         public abstract void DrawTriggerOptions(Vector2 size, float padX, float padY);
-
-        protected static bool GetResult(
-            float dataValue,
-            TriggerDataOp op,
-            float value)
-        {
-            return op switch
-            {
-                TriggerDataOp.Equals => dataValue == value,
-                TriggerDataOp.NotEquals => dataValue != value,
-                TriggerDataOp.LessThan => dataValue < value,
-                TriggerDataOp.GreaterThan => dataValue > value,
-                TriggerDataOp.LessThanEq => dataValue <= value,
-                TriggerDataOp.GreaterThanEq => dataValue >= value,
-                _ => false
-            };
-        }
     }
 }